Improve migration output messages with checkmarks and warnings for better visibility. Implement migrate rollback using down method to delete tables from db

+++ Migrations/20250115_01_create_categories_table.php
@@ -21,7 +21,7 @@ class CreateCategoriesTable {
         $result = $this->conn->query($sql);
         // Check if result is not empty
         if($result->rowCount() > 0) {
-            echo "\033[33m - categories table already exists\033[0m\n";
+            echo "\033[33m ! - categories table already exists\033[0m\n";
             return;
         }
 
@@ -32,7 +32,7 @@ class CreateCategoriesTable {
             created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP
         )";
         $this->conn->exec($sql);
-        echo " - Table categories created successfully          \033[32mDONE\033[0m\n";
+        echo "\033[32m \u{2714} DONE\033[0m       - Table categories created successfully.\n";
     }
 
     /**
@@ -40,9 +40,18 @@ class CreateCategoriesTable {
      */
     public function down() {
 
+        // Check if the table does not exist
+        $sql = "SHOW TABLES LIKE 'categories'";
+        $result = $this->conn->query($sql);
+        // Check if result is empty
+        if($result->rowCount() == 0) {
+            echo "\033[33m ! - categories table does not exist\033[0m\n";
+            return;
+        }
+
         $sql = "DROP TABLE categories";
         $this->conn->exec($sql);
-        echo "Table categories dropped successfully         \033[31mDONE\033[0m\n";
+        echo "\033[32m \u{2714} DONE\033[0m       - Table categories dropped successfully.\n";
     }
 }
 ?>
