+++ Migrations/20250115_04_attributes_table.php
@@ -23,7 +23,7 @@ class CreatAttributesTable {
         $result = $this->conn->query($sql);
         // Check if result is not empty
         if($result->rowCount() > 0) {
-            echo "\033[33m - attributes table already exists\033[0m\n";
+            echo "\033[33m ! - attributes table already exists\033[0m\n";
             return;
         }
         // Create the table
@@ -35,16 +35,25 @@ class CreatAttributesTable {
             created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP
         )";
         $this->conn->exec($sql);
-        echo " - Table attributes created successfully          \033[32mDONE\033[0m\n";
+         echo "\033[32m \u{2714} DONE\033[0m       - Table attributes created successfully.\n";
     }
 
     /**
      * Reverse the migrations
      */
     public function down() {
-        
+
+        // Check if the table does not exist
+        $sql = "SHOW TABLES LIKE 'attributes'";
+        $result = $this->conn->query($sql);
+        // Check if result is empty
+        if($result->rowCount() == 0) {
+            echo "\033[33m ! - attributes table does not exist\033[0m\n";
+            return;
+        }
+
         $sql = "DROP TABLE attributes";
         $this->conn->exec($sql);
-        echo "Table attributes dropped successfully         \033[31mDONE\033[0m\n";
+        echo "\033[32m \u{2714} DONE\033[0m       - Table attributes dropped successfully.\n";
     }
 }
