+++ Migrations/20250115_05_create_attribute_product_table.php
@@ -22,7 +22,7 @@ class CreateAttributeProductTable {
         $result = $this->conn->query($sql);
         // Check if result is not empty
         if($result->rowCount() > 0) {
-            echo "\033[33m - attribute_product table already exists\033[0m\n";
+            echo "\033[33m ! - attribute_product table already exists\033[0m\n";
             return;
         }
         // Create the table with foreign keys
@@ -36,16 +36,25 @@ class CreateAttributeProductTable {
             FOREIGN KEY (attribute_id) REFERENCES attributes(id)
         )";
         $this->conn->exec($sql);
-        echo " - Table attribute_product created successfully          \033[32mDONE\033[0m\n";
+         echo "\033[32m \u{2714} DONE\033[0m       - Table attribute_product created successfully.\n";
     }
 
     /**
      * Reverse the migrations
      */
     public function down() {
-        
+
+        // Check if the table does not exist
+        $sql = "SHOW TABLES LIKE 'attribute_product'";
+        $result = $this->conn->query($sql);
+        // Check if result is empty
+        if($result->rowCount() == 0) {
+            echo "\033[33m ! - attribute_product table does not exist\033[0m\n";
+            return;
+        }
+
         $sql = "DROP TABLE attribute_product";
         $this->conn->exec($sql);
-        echo "Table attribute_product dropped successfully         \033[31mDONE\033[0m\n";
+        echo "\033[32m \u{2714} DONE\033[0m       - Table attribute_product dropped successfully.\n";
     }
 }
