+++ Migrations/20250115_08_create_currency_price_table.php
@@ -22,7 +22,7 @@ class CreateCurrencyPriceTable {
         $result = $this->conn->query($sql);
         // Check if result is not empty
         if($result->rowCount() > 0) {
-            echo "\033[33m - currency_price table already exists\033[0m\n";
+            echo "\033[33m ! - currency_price table already exists\033[0m\n";
             return;
         }
         // Create the table
@@ -36,16 +36,25 @@ class CreateCurrencyPriceTable {
             FOREIGN KEY (currency_id) REFERENCES currency(id)
         )";
         $this->conn->exec($sql);
-        echo " - Table currency_price created successfully          \033[32mDONE\033[0m\n";
+         echo "\033[32m \u{2714} DONE\033[0m       - Table currency_price created successfully.\n";
     }
 
     /**
      * Reverse the migrations
      */
     public function down() {
-        
+
+        // Check if the table does not exist
+        $sql = "SHOW TABLES LIKE 'currency_price'";
+        $result = $this->conn->query($sql);
+        // Check if result is empty
+        if($result->rowCount() == 0) {
+            echo "\033[33m ! - currency_price table does not exist\033[0m\n";
+            return;
+        }
+
         $sql = "DROP TABLE currency_price";
         $this->conn->exec($sql);
-        echo "Table currency_price dropped successfully         \033[31mDONE\033[0m\n";
+        echo "\033[32m \u{2714} DONE\033[0m       - Table currency_price dropped successfully.\n";
     }
 }
